requirements.txt
langgraph
langchain
langchain-openai
python-dotenv
scikit-learn
psycopg2-binary
pandas
plotly
kaleido
tabulate